# Fboemer/faster decrypt