@fionser 2021-06-29
1. A quick question, why write explicitly out the tile = 1024 would give a better performance than a simple for-loop ? 
   Or why the compiler would not optimize this for us :(.
 2. Another question. Do you try to run this program on the same machine on parallel. 
     I mean when multiple SEAL's programs run on the same time, should the L1 Cache still in-and-out for these different program ?