@WeiDaiWD 2021-06-29
Thanks, Fabian! We are developing version 4.0.0 based on the BGV PR from Alibaba this summer. The progress is slow, like continental drift. :) For now, I'm restraining from releasing patches unless there is serious bug. PRs will be merged with more delay than usual. Please be patient with me.