@fboemer 2021-08-03
> 1. A quick question, why write explicitly out the tile = 1024 would give a better performance than a simple for-loop ?
>    Or why the compiler would not optimize this for us :(.
> 2. Another question. Do you try to run this program on the same machine on parallel.
>    I mean when multiple SEAL's programs run on the same time, should the L1 Cache still in-and-out for these different program ?

@fiosner, apologies for the delay on this.
1) Interestingly, I'm seeing some varying speedup/slowdown with the tiling approach, so I've removed the tiling for now.
![Screen Shot 2021-08-03 at 12 16 12 PM](https://user-images.githubusercontent.com/7935794/128073033-cc41054f-b4ee-4c43-8934-6da7ac10604a.png)

2) I've only tested single-threaded. To my knowledge, the L1 cache is unique to each core, so there should be no L1 cache interference between  different threads of SEALon different CPUs. In general, I would expect the L2 cache could create some slowdowns due to contention.