@WeiDaiWD 2021-09-15
Since this PR is made to contrib branch that includes BGV code, I cannot merge it with 3.6.6 easily. I applied your changes and named the commit as a pull request merge. So your commits won't be in the commit history; but your id and fork/branch/pr name will. I hope that it is fine with you.